Remove tools/hyper_hello
