+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
   "cli",
   "core",
-  "tools/hyper_hello",
   "deno_typescript",
   "test_plugin",
   "test_util",
