--- tools/hyper_hello/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "hyper_hello"
-version = "0.0.1"
-edition = "2018"
-
-[dependencies]
-hyper = "0.13.6"
-tokio = { version = "0.2.21", features = ["full"] }
-
-[[bin]]
-name = "hyper_hello"
-path = "hyper_hello.rs"
